pyproject.toml
[tool.poetry]
name = "nba-data"
version = "0.1.0"
description = ""
authors = ["Aumit Leon <aumitleon@gmail.com>"]
readme = "README.md"
packages = [{include = "nba_data"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^1.5.2"
duckdb = "^0.6.1"
google-cloud-storage = "^2.7.0"
httpx = "^0.23.1"
tenacity = "^8.1.0"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.19.4"
ipython = "^8.7.0"
ruff = "^0.0.199"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
